app/requirements.txt
Django==2.1.5
djangorestframework==3.9.1
entrypoints==0.3
flake8==3.7.1
mccabe==0.6.1
psycopg2==2.7.7
pycodestyle==2.5.0
pyflakes==2.1.0
pytz==2018.9
virtualenv==16.3.0
virtualenvwrapper-win==1.2.5
